src/main/webapp/app/entities/vente-produit-chad/route/vente-produit-chad-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IVenteProduitChad } from '../vente-produit-chad.model';
import { VenteProduitChadService } from '../service/vente-produit-chad.service';

import { VenteProduitChadRoutingResolveService } from './vente-produit-chad-routing-resolve.service';

describe('VenteProduitChad routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: VenteProduitChadRoutingResolveService;
  let service: VenteProduitChadService;
  let resultVenteProduitChad: IVenteProduitChad | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(VenteProduitChadRoutingResolveService);
    service = TestBed.inject(VenteProduitChadService);
    resultVenteProduitChad = undefined;
  });

  describe('resolve', () => {
    it('should return IVenteProduitChad returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultVenteProduitChad = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultVenteProduitChad).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultVenteProduitChad = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultVenteProduitChad).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IVenteProduitChad>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultVenteProduitChad = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultVenteProduitChad).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
